# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Plant Ride API",
    "description": "In-plant cab booking, pooling, fleet compliance and cost-centre billing for LHS.\n\n1. Call **POST /api/auth/login** with a demo account (password `Plant@123`):\n   `LHS-40218` (rider), `DRV-2281` (driver), `admin` (desk + admin), `desk` (desk).\n2. Click **Authorize** and paste the returned `token`.\n",
    "contact": {
      "name": "SRM ECO TECH",
      "email": "info@srmecotech.com"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/api/admin/routes/{id}/stops": {
      "put": {
        "tags": [
          "6. Admin"
        ],
        "summary": "Replace the stop sequence (refused while the route has live rides)",
        "operationId": "resequence",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResequenceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RouteDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/routes/{id}/rules": {
      "put": {
        "tags": [
          "6. Admin"
        ],
        "summary": "Change the occupancy cap, wait, detour, speed limit or frequency (audited, no release)",
        "operationId": "updateRules",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRouteRulesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RouteDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/desk/vehicles/{id}/status": {
      "post": {
        "tags": [
          "5. Transport desk"
        ],
        "summary": "Take a vehicle off-road (its riders are re-matched) or put it back on the road",
        "operationId": "vehicleStatus",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VehicleStatusRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/desk/bookings/{id}/reject": {
      "post": {
        "tags": [
          "5. Transport desk"
        ],
        "summary": "Reject an exclusive ride",
        "operationId": "reject",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReasonRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DeskActionResultDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/desk/bookings/{id}/cancel": {
      "post": {
        "tags": [
          "5. Transport desk"
        ],
        "summary": "Cancel a booking on the rider's behalf",
        "operationId": "cancel_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReasonRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DeskActionResultDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/desk/bookings/{id}/assign": {
      "post": {
        "tags": [
          "5. Transport desk"
        ],
        "summary": "Override: put a booking on a specific cab (hard rules and the cap still apply)",
        "operationId": "assign",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DeskActionResultDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/desk/bookings/{id}/approve": {
      "post": {
        "tags": [
          "5. Transport desk"
        ],
        "summary": "Approve an exclusive ride",
        "operationId": "approve",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DeskActionResultDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/desk/alerts/{id}/resolve": {
      "post": {
        "tags": [
          "5. Transport desk"
        ],
        "summary": "Resolve an alert",
        "operationId": "resolve",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AlertDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/desk/alerts/{id}/ack": {
      "post": {
        "tags": [
          "5. Transport desk"
        ],
        "summary": "Acknowledge an alert",
        "operationId": "acknowledge",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AlertDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/desk/bookings/{id}/candidates": {
      "get": {
        "tags": [
          "5. Transport desk"
        ],
        "summary": "Why each on-duty cab can or cannot take this booking (the engine's audit trail)",
        "operationId": "candidates",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CandidateDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/billing/summary": {
      "get": {
        "tags": [
          "6. Billing (admin)"
        ],
        "summary": "Charges by cost centre, vendor bill vs GPS, and compliance watch for a month",
        "operationId": "summary",
        "parameters": [
          {
            "name": "period",
            "in": "query",
            "description": "Billing month yyyy-MM; defaults to the current month",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BillingSummaryDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/audit": {
      "get": {
        "tags": [
          "6. Admin"
        ],
        "summary": "Who booked, changed, approved and overrode what",
        "operationId": "audit",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 50,
              "maximum": 200,
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuditLogDto"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ResequenceRequest": {
        "type": "object",
        "properties": {
          "version": {
            "type": "integer",
            "format": "int64"
          },
          "stops": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RouteStopInput"
            },
            "maxItems": 30,
            "minItems": 2
          }
        },
        "required": [
          "stops",
          "version"
        ]
      },
      "RouteStopInput": {
        "type": "object",
        "properties": {
          "stopId": {
            "type": "integer",
            "format": "int64"
          },
          "legMinutes": {
            "type": "integer",
            "format": "int32",
            "maximum": 120,
            "minimum": 0
          },
          "legKm": {
            "type": "number",
            "maximum": 50.0,
            "minimum": 0.0
          }
        },
        "required": [
          "legKm",
          "legMinutes",
          "stopId"
        ]
      },
      "RouteDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "routeType": {
            "type": "string",
            "enum": [
              "LOOP",
              "ONE_WAY"
            ]
          },
          "serviceNote": {
            "type": "string"
          },
          "frequencyMin": {
            "type": "integer",
            "format": "int32"
          },
          "maxPassengers": {
            "type": "integer",
            "format": "int32"
          },
          "maxWaitMin": {
            "type": "integer",
            "format": "int32"
          },
          "maxDetourMin": {
            "type": "integer",
            "format": "int32"
          },
          "speedLimitKmh": {
            "type": "integer",
            "format": "int32"
          },
          "active": {
            "type": "boolean"
          },
          "version": {
            "type": "integer",
            "format": "int64"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "totalMinutes": {
            "type": "integer",
            "format": "int32"
          },
          "totalKm": {
            "type": "number"
          },
          "stops": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RouteStopDto"
            }
          }
        }
      },
      "RouteStopDto": {
        "type": "object",
        "properties": {
          "seq": {
            "type": "integer",
            "format": "int32"
          },
          "stopId": {
            "type": "integer",
            "format": "int64"
          },
          "stopCode": {
            "type": "string"
          },
          "stopName": {
            "type": "string"
          },
          "legMinutes": {
            "type": "integer",
            "format": "int32"
          },
          "legKm": {
            "type": "number"
          },
          "geofenceM": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "UpdateRouteRulesRequest": {
        "type": "object",
        "properties": {
          "maxPassengers": {
            "type": "integer",
            "format": "int32",
            "maximum": 12,
            "minimum": 1
          },
          "maxWaitMin": {
            "type": "integer",
            "format": "int32",
            "maximum": 30,
            "minimum": 1
          },
          "maxDetourMin": {
            "type": "integer",
            "format": "int32",
            "maximum": 60,
            "minimum": 0
          },
          "speedLimitKmh": {
            "type": "integer",
            "format": "int32",
            "maximum": 80,
            "minimum": 5
          },
          "frequencyMin": {
            "type": "integer",
            "format": "int32",
            "maximum": 240,
            "minimum": 5
          },
          "version": {
            "type": "integer",
            "format": "int64"
          }
        },
        "required": [
          "maxDetourMin",
          "maxPassengers",
          "maxWaitMin",
          "speedLimitKmh",
          "version"
        ]
      },
...
      "VehicleStatusRequest": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "OFF_ROAD"
            ]
          },
          "note": {
            "type": "string",
            "maxLength": 200,
            "minLength": 0
          }
        },
        "required": [
          "status"
        ]
      },
      "ReasonRequest": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "maxLength": 200,
            "minLength": 0
          }
        },
        "required": [
          "reason"
        ]
      },
      "DeskActionResultDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "booking": {
            "$ref": "#/components/schemas/QueueItemDto"
          }
        }
      },
      "QueueItemDto": {
        "type": "object",
        "properties": {
          "bookingId": {
            "type": "integer",
            "format": "int64"
          },
          "bookingCode": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "SCHEDULED",
              "PENDING_APPROVAL",
              "REQUESTED",
              "ASSIGNED",
              "ONBOARD",
              "COMPLETED",
              "CANCELLED",
              "NO_SHOW",
              "REJECTED"
            ]
          },
          "rideType": {
            "type": "string",
            "enum": [
              "SHARED",
              "EXCLUSIVE"
            ]
          },
          "riderName": {
            "type": "string"
          },
          "personnelNo": {
            "type": "string"
          },
          "costCentreCode": {
            "type": "string"
          },
          "routeCode": {
            "type": "string"
          },
          "fromStopName": {
            "type": "string"
          },
          "toStopName": {
            "type": "string"
          },
          "seats": {
            "type": "integer",
            "format": "int32"
          },
          "scheduledAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "waitingMinutes": {
            "type": "integer",
            "format": "int64"
          },
          "vehicleCode": {
            "type": "string"
          },
          "note": {
            "type": "string"
          }
        }
      },
      "AssignRequest": {
        "type": "object",
        "properties": {
          "vehicleId": {
            "type": "integer",
            "format": "int64"
          }
        },
        "required": [
          "vehicleId"
        ]
      },
      "AlertDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "type": {
            "type": "string",
            "enum": [
              "SOS",
              "OVERSPEED",
              "BREAKDOWN",
              "ACCIDENT",
              "DELAY",
              "FUEL",
              "DOC_EXPIRY",
              "NO_SHOW",
              "DEMAND",
              "ROUTE_DEVIATION",
              "OTHER"
            ]
          },
          "severity": {
            "type": "string",
            "enum": [
              "CRITICAL",
              "HIGH",
              "MEDIUM",
              "LOW"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "OPEN",
              "ACKNOWLEDGED",
              "RESOLVED"
            ]
          },
          "message": {
            "type": "string"
          },
          "vehicleCode": {
            "type": "string"
          },
          "bookingCode": {
            "type": "string"
          },
          "raisedBy": {
            "type": "string"
          },
          "latitude": {
            "type": "number",
            "format": "double"
          },
          "longitude": {
            "type": "number",
            "format": "double"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "minutesAgo": {
            "type": "integer",
            "format": "int64"
          },
          "acknowledgedBy": {
            "type": "string"
          },
          "resolvedBy": {
            "type": "string"
          }
        }
      },
...
      "CandidateDto": {
        "type": "object",
        "properties": {
          "vehicleId": {
            "type": "integer",
            "format": "int64"
          },
          "vehicleCode": {
            "type": "string"
          },
          "feasible": {
            "type": "boolean"
          },
          "etaMinutes": {
            "type": "integer",
            "format": "int32"
          },
          "seatsTaken": {
            "type": "integer",
            "format": "int32"
          },
          "capacity": {
            "type": "integer",
            "format": "int32"
          },
          "score": {
            "type": "number",
            "format": "double"
          },
          "reason": {
            "type": "string"
          }
        }
      },
...
      "BillingSummaryDto": {
        "type": "object",
        "properties": {
          "period": {
            "type": "string"
          },
          "draft": {
            "type": "boolean"
          },
          "sharingRule": {
            "type": "string"
          },
          "totals": {
            "$ref": "#/components/schemas/Totals"
          },
          "costCentres": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CostCentreCharge"
            }
          },
          "vendors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VendorReconciliation"
            }
          },
          "compliance": {
            "$ref": "#/components/schemas/ComplianceWatch"
          }
        }
      },
      "ComplianceWatch": {
        "type": "object",
        "properties": {
          "documentsExpiringSoon": {
            "type": "integer",
            "format": "int32"
          },
          "expiredDocuments": {
            "type": "integer",
            "format": "int32"
          },
          "blockedVehicles": {
            "type": "integer",
            "format": "int32"
          },
          "blockedDrivers": {
            "type": "integer",
            "format": "int32"
          },
          "fullyCompliantVehicles": {
            "type": "integer",
            "format": "int32"
          },
          "totalVehicles": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "CostCentreCharge": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "trips": {
            "type": "integer",
            "format": "int64"
          },
          "rides": {
            "type": "integer",
            "format": "int64"
          },
          "km": {
            "type": "number"
          },
          "amount": {
            "type": "number"
          },
          "budget": {
            "type": "number"
          },
          "budgetUsedPct": {
            "type": "integer",
            "format": "int32"
          },
          "budgetStatus": {
            "type": "string"
          }
        }
      },
      "Totals": {
        "type": "object",
        "properties": {
          "trips": {
            "type": "integer",
            "format": "int64"
          },
          "rides": {
            "type": "integer",
            "format": "int64"
          },
          "gpsKm": {
            "type": "number"
          },
          "amount": {
            "type": "number"
          },
          "ridersPerTrip": {
            "type": "number"
          }
        }
      },
      "VendorReconciliation": {
        "type": "object",
        "properties": {
          "vendorCode": {
            "type": "string"
          },
          "vendorName": {
            "type": "string"
          },
          "claimedKm": {
            "type": "number"
          },
          "gpsKm": {
            "type": "number"
          },
          "gapKm": {
            "type": "number"
          },
          "gapPct": {
            "type": "number"
          },
          "claimedAmount": {
            "type": "number"
          },
          "gpsAmount": {
            "type": "number"
          },
          "status": {
            "type": "string"
          }
        }
      },
...
      "AuditLogDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "actorName": {
            "type": "string"
          },
          "action": {
            "type": "string"
          },
          "entityType": {
            "type": "string"
          },
          "entityId": {
            "type": "integer",
            "format": "int64"
          },
          "details": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}